User validity checks on API delete requests

+++ CSGOTrackerAPI/CSGOTrackerAPI/Controllers/GamesController.cs
@@ -123,11 +123,20 @@ namespace CSGOTrackerAPI.Controllers
         [HttpDelete("{id}")]
         public async Task<ActionResult> DeleteGame(int id)
         {
-            var game = await _context.Games.FindAsync(id);
+            var user = await userManager.FindByNameAsync(User.Identity.Name);
+            if (user == null)
+            {
+                return Unauthorized();
+            }
 
+            var game = await _context.Games.FindAsync(id);
             if (game == null)
             {
                 return NotFound();
+            }
+            if (game.UserId != user.Id)
+            {
+                return Unauthorized();
             }
 
             _context.Games.Remove(game);
